Revert "chore: move server code to different files (#8223)"
This reverts commit abc705f033acfd4a25c7b54e74d69db9ca9022e5.
