+++ backend/server/server.go
@@ -10,13 +10,18 @@ import (
 	"sync"
 	"time"
 
+	"github.com/google/uuid"
 	grpcRuntime "github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/labstack/echo/v4"
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/reflection"
 	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/encoding/protojson"
+
+	v1pb "github.com/bytebase/bytebase/proto/generated-go/v1"
 
 	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/recovery"
 
@@ -31,7 +36,10 @@ import (
 	enterpriseAPI "github.com/bytebase/bytebase/backend/enterprise/api"
 	enterpriseService "github.com/bytebase/bytebase/backend/enterprise/service"
 	api "github.com/bytebase/bytebase/backend/legacyapi"
+	"github.com/bytebase/bytebase/backend/metric"
+	metricCollector "github.com/bytebase/bytebase/backend/metric/collector"
 	"github.com/bytebase/bytebase/backend/migrator"
+	"github.com/bytebase/bytebase/backend/plugin/app/feishu"
 	bbs3 "github.com/bytebase/bytebase/backend/plugin/storage/s3"
 	"github.com/bytebase/bytebase/backend/resources/mongoutil"
 	"github.com/bytebase/bytebase/backend/resources/mysqlutil"
@@ -48,6 +56,7 @@ import (
 	"github.com/bytebase/bytebase/backend/runner/taskrun"
 	"github.com/bytebase/bytebase/backend/store"
 	_ "github.com/bytebase/bytebase/docs/openapi" // initial the swagger doc
+	storepb "github.com/bytebase/bytebase/proto/generated-go/store"
 
 	// Register clickhouse driver.
 
@@ -158,7 +167,8 @@ type Server struct {
 	// Postgres utility binaries
 	pgBinDir string
 
-	s3Client *bbs3.Client
+	s3Client       *bbs3.Client
+	feishuProvider *feishu.Provider
 
 	// stateCfg is the shared in-momory state within the server.
 	stateCfg *state.State
@@ -286,11 +296,32 @@ func NewServer(ctx context.Context, profile config.Profile) (*Server, error) {
 	// Cache the license.
 	s.licenseService.LoadSubscription(ctx)
 
-	secret, externalURL, tokenDuration, err := s.getInitSetting(ctx, storeInstance)
+	config, err := s.getInitSetting(ctx, storeInstance)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to init config")
 	}
-	s.secret = secret
+	s.secret = config.secret
+
+	workspaceProfileSettingName := api.SettingWorkspaceProfile
+	setting, err := s.store.GetSettingV2(ctx, &store.FindSettingMessage{Name: &workspaceProfileSettingName})
+	if err != nil {
+		return nil, err
+	}
+	tokenDuration := auth.DefaultTokenDuration
+	externalURL := ""
+	if setting != nil {
+		settingValue := new(storepb.WorkspaceProfileSetting)
+		if err := protojson.Unmarshal([]byte(setting.Value), settingValue); err != nil {
+			return nil, err
+		}
+		if settingValue.ExternalUrl != "" {
+			externalURL = settingValue.ExternalUrl
+		}
+		if settingValue.TokenDuration != nil && settingValue.TokenDuration.Seconds > 0 {
+			tokenDuration = settingValue.TokenDuration.AsDuration()
+		}
+	}
+
 	s.activityManager = activity.NewManager(storeInstance)
 	s.dbFactory = dbfactory.New(s.mysqlBinDir, s.mongoBinDir, s.pgBinDir, profile.DataDir, s.secret)
 
@@ -325,10 +356,6 @@ func NewServer(ctx context.Context, profile config.Profile) (*Server, error) {
 	if !profile.Readonly {
 		s.slowQuerySyncer = slowquerysync.NewSyncer(storeInstance, s.dbFactory, s.stateCfg, profile)
 		s.backupRunner = backuprun.NewRunner(storeInstance, s.dbFactory, s.s3Client, s.stateCfg, &profile)
-		s.rollbackRunner = rollbackrun.NewRunner(&profile, storeInstance, s.dbFactory, s.stateCfg)
-		s.mailSender = mail.NewSender(s.store, s.stateCfg)
-		s.relayRunner = relay.NewRunner(storeInstance, s.activityManager, s.stateCfg)
-		s.approvalRunner = approval.NewRunner(storeInstance, s.dbFactory, s.stateCfg, s.activityManager, s.relayRunner, s.licenseService)
 
 		s.taskSchedulerV2 = taskrun.NewSchedulerV2(storeInstance, s.stateCfg, s.activityManager)
 		s.taskSchedulerV2.Register(api.TaskGeneral, taskrun.NewDefaultExecutor())
@@ -343,19 +370,26 @@ func NewServer(ctx context.Context, profile config.Profile) (*Server, error) {
 		s.taskSchedulerV2.Register(api.TaskDatabaseRestorePITRRestore, taskrun.NewPITRRestoreExecutor(storeInstance, s.dbFactory, s.s3Client, s.schemaSyncer, s.stateCfg, profile))
 		s.taskSchedulerV2.Register(api.TaskDatabaseRestorePITRCutover, taskrun.NewPITRCutoverExecutor(storeInstance, s.dbFactory, s.schemaSyncer, s.backupRunner, s.activityManager, profile))
 
-		s.planCheckScheduler = plancheck.NewScheduler(storeInstance, s.licenseService, s.stateCfg)
-		databaseConnectExecutor := plancheck.NewDatabaseConnectExecutor(storeInstance, s.dbFactory)
-		s.planCheckScheduler.Register(store.PlanCheckDatabaseConnect, databaseConnectExecutor)
-		statementTypeExecutor := plancheck.NewStatementTypeExecutor(storeInstance, s.dbFactory)
-		s.planCheckScheduler.Register(store.PlanCheckDatabaseStatementType, statementTypeExecutor)
-		statementAdviseExecutor := plancheck.NewStatementAdviseExecutor(storeInstance, s.dbFactory, s.licenseService)
-		s.planCheckScheduler.Register(store.PlanCheckDatabaseStatementAdvise, statementAdviseExecutor)
-		ghostSyncExecutor := plancheck.NewGhostSyncExecutor(storeInstance, s.secret)
-		s.planCheckScheduler.Register(store.PlanCheckDatabaseGhostSync, ghostSyncExecutor)
-		pitrMySQLExecutor := plancheck.NewPITRMySQLExecutor(storeInstance, s.dbFactory)
-		s.planCheckScheduler.Register(store.PlanCheckDatabasePITRMySQL, pitrMySQLExecutor)
-		statementReportExecutor := plancheck.NewStatementReportExecutor(storeInstance, s.dbFactory)
-		s.planCheckScheduler.Register(store.PlanCheckDatabaseStatementSummaryReport, statementReportExecutor)
+		s.rollbackRunner = rollbackrun.NewRunner(&profile, storeInstance, s.dbFactory, s.stateCfg)
+		s.mailSender = mail.NewSender(s.store, s.stateCfg)
+		s.relayRunner = relay.NewRunner(storeInstance, s.activityManager, s.stateCfg)
+		s.approvalRunner = approval.NewRunner(storeInstance, s.dbFactory, s.stateCfg, s.activityManager, s.relayRunner, s.licenseService)
+
+		{
+			s.planCheckScheduler = plancheck.NewScheduler(storeInstance, s.licenseService, s.stateCfg)
+			databaseConnectExecutor := plancheck.NewDatabaseConnectExecutor(storeInstance, s.dbFactory)
+			s.planCheckScheduler.Register(store.PlanCheckDatabaseConnect, databaseConnectExecutor)
+			statementTypeExecutor := plancheck.NewStatementTypeExecutor(storeInstance, s.dbFactory)
+			s.planCheckScheduler.Register(store.PlanCheckDatabaseStatementType, statementTypeExecutor)
+			statementAdviseExecutor := plancheck.NewStatementAdviseExecutor(storeInstance, s.dbFactory, s.licenseService)
+			s.planCheckScheduler.Register(store.PlanCheckDatabaseStatementAdvise, statementAdviseExecutor)
+			ghostSyncExecutor := plancheck.NewGhostSyncExecutor(storeInstance, s.secret)
+			s.planCheckScheduler.Register(store.PlanCheckDatabaseGhostSync, ghostSyncExecutor)
+			pitrMySQLExecutor := plancheck.NewPITRMySQLExecutor(storeInstance, s.dbFactory)
+			s.planCheckScheduler.Register(store.PlanCheckDatabasePITRMySQL, pitrMySQLExecutor)
+			statementReportExecutor := plancheck.NewStatementReportExecutor(storeInstance, s.dbFactory)
+			s.planCheckScheduler.Register(store.PlanCheckDatabaseStatementSummaryReport, statementReportExecutor)
+		}
 
 		// Metric reporter
 		s.initMetricReporter()
@@ -394,30 +428,330 @@ func NewServer(ctx context.Context, profile config.Profile) (*Server, error) {
 		),
 	)
 	configureEchoRouters(s.e, s.grpcServer, mux)
-	postCreateUser := func(ctx context.Context, user *store.UserMessage, firstEndUser bool) error {
-		if profile.TestOnlySkipOnboardingData {
-			return nil
-		}
-		// Only generate onboarding data after the first enduser signup.
-		if firstEndUser {
-			if profile.SampleDatabasePort != 0 {
-				if err := s.generateOnboardingData(ctx, user); err != nil {
-					return status.Errorf(codes.Internal, "failed to prepare onboarding data, error: %v", err)
+
+	v1pb.RegisterAuthServiceServer(s.grpcServer, v1.NewAuthService(s.store, s.secret, tokenDuration, s.licenseService, s.metricReporter, &profile,
+		func(ctx context.Context, user *store.UserMessage, firstEndUser bool) error {
+			if s.profile.TestOnlySkipOnboardingData {
+				return nil
+			}
+			// Only generate onboarding data after the first enduser signup.
+			if firstEndUser {
+				if profile.SampleDatabasePort != 0 {
+					if err := s.generateOnboardingData(ctx, user); err != nil {
+						return status.Errorf(codes.Internal, "failed to prepare onboarding data, error: %v", err)
+					}
 				}
 			}
-		}
-		return nil
-	}
-	rolloutService, issueService, err := configureGrpcRouters(ctx, mux, s.grpcServer, s.store, s.dbFactory, s.licenseService, &s.profile, s.metricReporter, s.stateCfg, s.schemaSyncer, s.activityManager, s.backupRunner, s.relayRunner, s.planCheckScheduler, postCreateUser, s.secret, &s.errorRecordRing, tokenDuration)
+			return nil
+		}))
+	v1pb.RegisterActuatorServiceServer(s.grpcServer, v1.NewActuatorService(s.store, &s.profile, &s.errorRecordRing))
+	v1pb.RegisterSubscriptionServiceServer(s.grpcServer, v1.NewSubscriptionService(
+		s.store,
+		&s.profile,
+		s.metricReporter,
+		s.licenseService))
+	v1pb.RegisterEnvironmentServiceServer(s.grpcServer, v1.NewEnvironmentService(s.store, s.licenseService))
+	v1pb.RegisterInstanceServiceServer(s.grpcServer, v1.NewInstanceService(
+		s.store,
+		s.licenseService,
+		s.metricReporter,
+		s.secret,
+		s.stateCfg,
+		s.dbFactory,
+		s.schemaSyncer))
+	v1pb.RegisterProjectServiceServer(s.grpcServer, v1.NewProjectService(s.store, s.activityManager, s.licenseService))
+	v1pb.RegisterDatabaseServiceServer(s.grpcServer, v1.NewDatabaseService(s.store, s.backupRunner, s.schemaSyncer, s.licenseService))
+	v1pb.RegisterInstanceRoleServiceServer(s.grpcServer, v1.NewInstanceRoleService(s.store, s.dbFactory))
+	v1pb.RegisterOrgPolicyServiceServer(s.grpcServer, v1.NewOrgPolicyService(s.store, s.licenseService))
+	v1pb.RegisterIdentityProviderServiceServer(s.grpcServer, v1.NewIdentityProviderService(s.store, s.licenseService))
+	v1pb.RegisterSettingServiceServer(s.grpcServer, v1.NewSettingService(s.store, &s.profile, s.licenseService, s.stateCfg, s.feishuProvider))
+	v1pb.RegisterAnomalyServiceServer(s.grpcServer, v1.NewAnomalyService(s.store))
+	v1pb.RegisterSQLServiceServer(s.grpcServer, v1.NewSQLService(s.store, s.schemaSyncer, s.dbFactory, s.activityManager, s.licenseService))
+	v1pb.RegisterExternalVersionControlServiceServer(s.grpcServer, v1.NewExternalVersionControlService(s.store))
+	v1pb.RegisterRiskServiceServer(s.grpcServer, v1.NewRiskService(s.store, s.licenseService))
+	s.issueService = v1.NewIssueService(s.store, s.activityManager, s.relayRunner, s.stateCfg, s.licenseService, s.metricReporter)
+	v1pb.RegisterIssueServiceServer(s.grpcServer, s.issueService)
+	s.rolloutService = v1.NewRolloutService(s.store, s.licenseService, s.dbFactory, s.planCheckScheduler, s.stateCfg, s.activityManager)
+	v1pb.RegisterRolloutServiceServer(s.grpcServer, s.rolloutService)
+	v1pb.RegisterRoleServiceServer(s.grpcServer, v1.NewRoleService(s.store, s.licenseService))
+	v1pb.RegisterSheetServiceServer(s.grpcServer, v1.NewSheetService(s.store, s.licenseService))
+	v1pb.RegisterSchemaDesignServiceServer(s.grpcServer, v1.NewSchemaDesignService(s.store, s.licenseService))
+	v1pb.RegisterCelServiceServer(s.grpcServer, v1.NewCelService())
+	v1pb.RegisterLoggingServiceServer(s.grpcServer, v1.NewLoggingService(s.store))
+	v1pb.RegisterBookmarkServiceServer(s.grpcServer, v1.NewBookmarkService(s.store))
+	v1pb.RegisterInboxServiceServer(s.grpcServer, v1.NewInboxService(s.store))
+	reflection.Register(s.grpcServer)
+
+	// REST gateway proxy.
+	grpcEndpoint := fmt.Sprintf(":%d", profile.GrpcPort)
+	grpcConn, err := grpc.Dial(grpcEndpoint, grpc.WithTransportCredentials(insecure.NewCredentials()))
 	if err != nil {
 		return nil, err
 	}
-	s.rolloutService, s.issueService = rolloutService, issueService
 
-	reflection.Register(s.grpcServer)
+	if err := v1pb.RegisterAuthServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterActuatorServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterSubscriptionServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterEnvironmentServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterInstanceServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterProjectServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterDatabaseServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterInstanceRoleServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterOrgPolicyServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterIdentityProviderServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterSettingServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterAnomalyServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterSQLServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterExternalVersionControlServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterRoleServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterSheetServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterRolloutServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterIssueServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterLoggingServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterBookmarkServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
+	if err := v1pb.RegisterInboxServiceHandler(ctx, mux, grpcConn); err != nil {
+		return nil, err
+	}
 
 	serverStarted = true
 	return s, nil
+}
+
+// initMetricReporter will initial the metric scheduler.
+func (s *Server) initMetricReporter() {
+	metricReporter := metricreport.NewReporter(s.store, s.licenseService, &s.profile, s.profile.EnableMetric)
+	metricReporter.Register(metric.InstanceCountMetricName, metricCollector.NewInstanceCountCollector(s.store))
+	metricReporter.Register(metric.IssueCountMetricName, metricCollector.NewIssueCountCollector(s.store))
+	metricReporter.Register(metric.ProjectCountMetricName, metricCollector.NewProjectCountCollector(s.store))
+	metricReporter.Register(metric.PolicyCountMetricName, metricCollector.NewPolicyCountCollector(s.store))
+	metricReporter.Register(metric.TaskCountMetricName, metricCollector.NewTaskCountCollector(s.store))
+	metricReporter.Register(metric.DatabaseCountMetricName, metricCollector.NewDatabaseCountCollector(s.store))
+	metricReporter.Register(metric.SheetCountMetricName, metricCollector.NewSheetCountCollector(s.store))
+	metricReporter.Register(metric.MemberCountMetricName, metricCollector.NewMemberCountCollector(s.store))
+	s.metricReporter = metricReporter
+}
+
+// retrieved via the SettingService upon startup.
+type workspaceConfig struct {
+	// secret used to sign the JWT auth token
+	secret string
+	// workspaceID used to initial the identify for a new workspace.
+	workspaceID string
+}
+
+func (s *Server) getInitSetting(ctx context.Context, datastore *store.Store) (*workspaceConfig, error) {
+	// secretLength is the length for the secret used to sign the JWT auto token.
+	const secretLength = 32
+
+	// initial branding
+	if _, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingBrandingLogo,
+		Value:       "",
+		Description: "The branding slogo image in base64 string format.",
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+
+	conf := &workspaceConfig{}
+
+	// initial JWT token
+	value, err := common.RandomString(secretLength)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to generate random JWT secret")
+	}
+	authSetting, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingAuthSecret,
+		Value:       value,
+		Description: "Random string used to sign the JWT auth token.",
+	}, api.SystemBotID)
+	if err != nil {
+		return nil, err
+	}
+	conf.secret = authSetting.Value
+
+	// initial workspace
+	workspaceSetting, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingWorkspaceID,
+		Value:       uuid.New().String(),
+		Description: "The workspace identifier",
+	}, api.SystemBotID)
+	if err != nil {
+		return nil, err
+	}
+	conf.workspaceID = workspaceSetting.Value
+
+	// initial license
+	if _, _, err = datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingEnterpriseLicense,
+		Value:       "",
+		Description: "Enterprise license",
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+
+	// initial feishu app
+	if _, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingAppIM,
+		Value:       "{}",
+		Description: "",
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+
+	// initial watermark setting
+	if _, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingWatermark,
+		Value:       "0",
+		Description: "Display watermark",
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+
+	// initial OpenAI key setting
+	if _, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingPluginOpenAIKey,
+		Value:       "",
+		Description: "API key to request OpenAI (ChatGPT)",
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+	if _, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingPluginOpenAIEndpoint,
+		Value:       "",
+		Description: "API Endpoint for OpenAI",
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+
+	// initial external approval setting
+	externalApprovalSettingValue, err := protojson.Marshal(&storepb.ExternalApprovalSetting{})
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to marshal initial external approval setting")
+	}
+	if _, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingWorkspaceExternalApproval,
+		Value:       string(externalApprovalSettingValue),
+		Description: "The external approval setting",
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+
+	// initial schema template setting
+	schemaTemplateSettingValue, err := protojson.Marshal(&storepb.SchemaTemplateSetting{})
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to marshal initial schema template setting")
+	}
+	if _, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingSchemaTemplate,
+		Value:       string(schemaTemplateSettingValue),
+		Description: "The schema template setting",
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+
+	// initial data classification setting
+	dataClassificationSettingValue, err := protojson.Marshal(&storepb.DataClassificationSetting{})
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to marshal initial data classification setting")
+	}
+	if _, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name:        api.SettingDataClassification,
+		Value:       string(dataClassificationSettingValue),
+		Description: "The data classification setting",
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+
+	// initial workspace approval setting
+	approvalSettingValue, err := protojson.Marshal(&storepb.WorkspaceApprovalSetting{})
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to marshal initial workspace approval setting")
+	}
+	if _, _, err := datastore.CreateSettingIfNotExistV2(ctx, &store.SettingMessage{
+		Name: api.SettingWorkspaceApproval,
+		// Value is ""
+		Value:       string(approvalSettingValue),
+		Description: "The workspace approval setting",
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+
+	// initial workspace profile setting
+	settingName := api.SettingWorkspaceProfile
+	workspaceProfileSetting, err := s.store.GetSettingV2(ctx, &store.FindSettingMessage{
+		Name:    &settingName,
+		Enforce: true,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	workspaceProfilePayload := &storepb.WorkspaceProfileSetting{
+		ExternalUrl: s.profile.ExternalURL,
+	}
+	if workspaceProfileSetting != nil {
+		workspaceProfilePayload = new(storepb.WorkspaceProfileSetting)
+		if err := protojson.Unmarshal([]byte(workspaceProfileSetting.Value), workspaceProfilePayload); err != nil {
+			return nil, err
+		}
+		if s.profile.ExternalURL != "" {
+			workspaceProfilePayload.ExternalUrl = s.profile.ExternalURL
+		}
+	}
+
+	bytes, err := protojson.Marshal(workspaceProfilePayload)
+	if err != nil {
+		return nil, err
+	}
+
+	if _, err := datastore.UpsertSettingV2(ctx, &store.SetSettingMessage{
+		Name:  api.SettingWorkspaceProfile,
+		Value: string(bytes),
+	}, api.SystemBotID); err != nil {
+		return nil, err
+	}
+
+	return conf, nil
 }
 
 // Run will run the server.
